backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Mail==0.9.1
pymongo==4.5.0
python-dotenv==1.0.0
Pillow==11.0.0
python-socketio==5.9.0
eventlet==0.40.3
bcrypt==4.0.1
PyJWT==2.8.0
requests==2.31.0
geopy==2.3.0
twilio==8.5.0
opencv-python==4.8.1.78
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.24.0
ultralytics==8.0.235
pyyaml==6.0.1
scipy==1.11.4
torch==2.1.1
torchvision==0.16.1